tryten/requirements.txt
appdirs==1.4.3
defusedxml==0.5.0
Django==1.10
django-allauth==0.32.0
django-crispy-forms==1.6.1
oauthlib==2.0.2
packaging==16.8
pyparsing==2.2.0
python3-openid==3.1.0
requests==2.14.2
requests-oauthlib==0.8.0
six==1.10.0
stripe==1.56.0
